HACK for parsing DISABLEd constraints
Postgres doesn't support DISABLE on a constraint, so we need to excise these.

My preferred option was to match DISABLE as a constraint-deferreables option
and then throw an exception (X::SkipStatement) to be caught somewhere up
higher, but I couldn't figure out how to catch an exception within a grammar.