Fix slowness when parsing constraints
The previous refactoring to change ';' parsing made constraint parsing super
slow.  This change speeds it up by more explicitly matching the broken
constraints we need to drop.  It also has extra logic in the 'ADD CONSTRAINT'
handler to know when the 'DISABLE' option is used, and so should drop the
line