Remove unnecessary rule/action
Previously, the plate_locations view wasn't parsable because it contained
an infix expression with a function on the left side of the operator,
while the rule was only looking for a identifier-or-value.  A kludge was
added to speficially look for this view definition and pass it through
unchanged.

The changes in this branch allow the view to parse properly, so we can remove
the special handling code.