Omit schema names prepending table aliases
Oracle apparently allows schema.alias as an identifier, but Postgres
does not.  Remember the aliases we see in FROM clauses, and if they
show up as entity-names later as schema.alias or schema.alias.column,
remove the schema.

Note that this solution only works for alias use that appears after
it's defined.  For the case where these aliases appear before they're
defined (for example, if they're in the SELECT clause), then we'll need
a different solution.